+++ src/main/java/lotto/domain/LottoReport.java
@@ -1,9 +1,6 @@
-package lotto;
+package lotto.domain;
 
 import lotto.domain.prize.LottoPrize;
-import lotto.domain.Lottos;
-import lotto.domain.PurchaseAmount;
-import lotto.domain.WinningLotto;
 import lotto.domain.prize.LottoReward;
 import lotto.view.ResultView;
