+++ src/main/java/lotto/view/ResultView.java
@@ -1,6 +1,6 @@
 package lotto.view;
 
-import lotto.LottoReport;
+import lotto.domain.LottoReport;
 import lotto.domain.prize.LottoPrize;
 import lotto.domain.Lottos;
