+++ src/test/java/lotto/view/ResultViewTest.java
@@ -6,7 +6,7 @@ import lotto.domain.Lotto;
 import lotto.domain.LottoNumber;
 import lotto.domain.prize.LottoPrize;
 import lotto.domain.Lottos;
-import lotto.LottoReport;
+import lotto.domain.LottoReport;
 import lotto.domain.PurchaseAmount;
 import lotto.domain.WinningLotto;
 import org.junit.jupiter.api.DisplayName;
